move tag to api package. fix import paths for tag
