+++ api/provider/aws/deploy.go
@@ -1,17 +1,17 @@
 package aws
 
 import (
+	"github.com/quintilesims/layer0/api/tag"
 	awsc "github.com/quintilesims/layer0/common/aws"
-	"github.com/quintilesims/layer0/common/db/tag_store"
 	"github.com/quintilesims/layer0/common/models"
 )
 
 type DeployProvider struct {
 	AWS      *awsc.Client
-	TagStore tag_store.TagStore
+	TagStore tag.Store
 }
 
-func NewDeployProvider(a *awsc.Client, t tag_store.TagStore) *DeployProvider {
+func NewDeployProvider(a *awsc.Client, t tag.Store) *DeployProvider {
 	return &DeployProvider{
 		AWS:      a,
 		TagStore: t,
