+++ api/provider/aws/environment.go
@@ -1,18 +1,18 @@
 package aws
 
 import (
+	"github.com/quintilesims/layer0/api/tag"
 	awsc "github.com/quintilesims/layer0/common/aws"
 	"github.com/quintilesims/layer0/common/config"
-	"github.com/quintilesims/layer0/common/db/tag_store"
 )
 
 type EnvironmentProvider struct {
 	AWS      *awsc.Client
-	TagStore tag_store.TagStore
+	TagStore tag.Store
 	Config   config.APIConfig
 }
 
-func NewEnvironmentProvider(a *awsc.Client, t tag_store.TagStore, c config.APIConfig) *EnvironmentProvider {
+func NewEnvironmentProvider(a *awsc.Client, t tag.Store, c config.APIConfig) *EnvironmentProvider {
 	return &EnvironmentProvider{
 		AWS:      a,
 		TagStore: t,
