+++ api/provider/aws/load_balancer.go
@@ -1,18 +1,18 @@
 package aws
 
 import (
+	"github.com/quintilesims/layer0/api/tag"
 	awsc "github.com/quintilesims/layer0/common/aws"
 	"github.com/quintilesims/layer0/common/config"
-	"github.com/quintilesims/layer0/common/db/tag_store"
 )
 
 type LoadBalancerProvider struct {
 	AWS      *awsc.Client
-	TagStore tag_store.TagStore
+	TagStore tag.Store
 	Config   config.APIConfig
 }
 
-func NewLoadBalancerProvider(a *awsc.Client, t tag_store.TagStore, c config.APIConfig) *LoadBalancerProvider {
+func NewLoadBalancerProvider(a *awsc.Client, t tag.Store, c config.APIConfig) *LoadBalancerProvider {
 	return &LoadBalancerProvider{
 		AWS:      a,
 		TagStore: t,
